Extract DataClassMatcher / StructClassMatcher mixins
`struct_like_class?` was duplicated verbatim across six cops (three for
`Data.define`, three for `Struct.new`), and the `Struct.new`-specific
`attr_argument?` override was duplicated in two of them.

Extract them into DataClassMatcher and StructClassMatcher under mixin/,
included by the respective cops. StructClassMatcher carries the
`attr_argument?` override, so cops that also include MissingClassAnnotation /
ClassCommentAlignment include it last to keep the override ahead of
DataStructHelper in the ancestor chain.
