+++ lib/rubocop/cop/rbs_inline_cops.rb
@@ -9,7 +9,9 @@ require_relative "style/rbs_inline/mixin/comment_parser"
 require_relative "style/rbs_inline/mixin/data_struct_helper"
 require_relative "style/rbs_inline/mixin/missing_class_annotation"
 require_relative "style/rbs_inline/mixin/class_comment_alignment"
+require_relative "style/rbs_inline/mixin/data_class_matcher"
 require_relative "style/rbs_inline/mixin/file_filter"
+require_relative "style/rbs_inline/mixin/struct_class_matcher"
 
 require_relative "style/rbs_inline/data_class_comment_alignment"
 require_relative "style/rbs_inline/data_define_with_block"
