+++ lib/rubocop/cop/style/rbs_inline/missing_struct_class_annotation.rb
@@ -24,6 +24,7 @@ module RuboCop
         class MissingStructClassAnnotation < Base
           prepend FileFilter
           include MissingClassAnnotation
+          include StructClassMatcher
           extend AutoCorrector
 
           MSG = "Missing inline type annotation for Struct attribute (e.g., `#: Type`)."
@@ -33,22 +34,6 @@ module RuboCop
             return unless struct_like_class?(node)
 
             check_missing_annotations(node)
-          end
-
-          private
-
-          # @rbs node: RuboCop::AST::SendNode
-          def struct_like_class?(node) #: bool
-            return false unless node.method_name == :new
-
-            (r = node.receiver).is_a?(RuboCop::AST::ConstNode) && r.short_name == :Struct
-          end
-
-          # `Struct.new` treats a leading string argument as the struct name, so only
-          # symbol arguments are attributes.
-          # @rbs arg: RuboCop::AST::Node
-          def attr_argument?(arg) #: bool
-            arg.sym_type?
           end
         end
       end
