+++ lib/rubocop/cop/style/rbs_inline/struct_class_comment_alignment.rb
@@ -26,6 +26,7 @@ module RuboCop
         class StructClassCommentAlignment < Base
           prepend FileFilter
           include ClassCommentAlignment
+          include StructClassMatcher
           extend AutoCorrector
 
           MSG = "Misaligned inline type annotation for Struct attribute."
@@ -35,22 +36,6 @@ module RuboCop
             return unless struct_like_class?(node)
 
             check_comment_alignment(node)
-          end
-
-          private
-
-          # @rbs node: RuboCop::AST::SendNode
-          def struct_like_class?(node) #: bool
-            return false unless node.method_name == :new
-
-            (r = node.receiver).is_a?(RuboCop::AST::ConstNode) && r.short_name == :Struct
-          end
-
-          # `Struct.new` treats a leading string argument as the struct name, so only
-          # symbol arguments are attributes.
-          # @rbs arg: RuboCop::AST::Node
-          def attr_argument?(arg) #: bool
-            arg.sym_type?
           end
         end
       end
