+++ sig/rubocop/cop/style/rbs_inline/data_class_comment_alignment.rbs
@@ -29,17 +29,14 @@ module RuboCop
 
           include ClassCommentAlignment
 
+          include DataClassMatcher
+
           extend AutoCorrector
 
           MSG: ::String
 
           # @rbs node: RuboCop::AST::SendNode
           def on_send: (RuboCop::AST::SendNode node) -> void
-
-          private
-
-          # @rbs node: RuboCop::AST::SendNode
-          def struct_like_class?: (RuboCop::AST::SendNode node) -> bool
         end
       end
     end
