+++ sig/rubocop/cop/style/rbs_inline/missing_data_class_annotation.rbs
@@ -24,17 +24,14 @@ module RuboCop
 
           include MissingClassAnnotation
 
+          include DataClassMatcher
+
           extend AutoCorrector
 
           MSG: ::String
 
           # @rbs node: RuboCop::AST::SendNode
           def on_send: (RuboCop::AST::SendNode node) -> void
-
-          private
-
-          # @rbs node: RuboCop::AST::SendNode
-          def struct_like_class?: (RuboCop::AST::SendNode node) -> bool
         end
       end
     end
