+++ sig/rubocop/cop/style/rbs_inline/mixin/struct_class_matcher.rbs
@@ -1,0 +1,26 @@
+# Generated from lib/rubocop/cop/style/rbs_inline/mixin/struct_class_matcher.rb with RBS::Inline
+
+module RuboCop
+  module Cop
+    module Style
+      module RbsInline
+        # Matcher for `Struct.new` calls, shared by the cops that check `Struct.new`
+        # class definitions.
+        #
+        # Also overrides {DataStructHelper#attr_argument?} for `Struct.new` semantics,
+        # so cops that include both must include this module last.
+        module StructClassMatcher
+          private
+
+          # @rbs node: RuboCop::AST::SendNode
+          def struct_like_class?: (RuboCop::AST::SendNode node) -> bool
+
+          # `Struct.new` treats a leading string argument as the struct name, so only
+          # symbol arguments are attributes.
+          # @rbs arg: RuboCop::AST::Node
+          def attr_argument?: (RuboCop::AST::Node arg) -> bool
+        end
+      end
+    end
+  end
+end
