+++ sig/rubocop/cop/style/rbs_inline/struct_class_comment_alignment.rbs
@@ -27,22 +27,14 @@ module RuboCop
 
           include ClassCommentAlignment
 
+          include StructClassMatcher
+
           extend AutoCorrector
 
           MSG: ::String
 
           # @rbs node: RuboCop::AST::SendNode
           def on_send: (RuboCop::AST::SendNode node) -> void
-
-          private
-
-          # @rbs node: RuboCop::AST::SendNode
-          def struct_like_class?: (RuboCop::AST::SendNode node) -> bool
-
-          # `Struct.new` treats a leading string argument as the struct name, so only
-          # symbol arguments are attributes.
-          # @rbs arg: RuboCop::AST::Node
-          def attr_argument?: (RuboCop::AST::Node arg) -> bool
         end
       end
     end
