+++ sig/rubocop/cop/style/rbs_inline/struct_new_with_block.rbs
@@ -26,15 +26,12 @@ module RuboCop
         class StructNewWithBlock < Base
           prepend FileFilter
 
+          include StructClassMatcher
+
           MSG: ::String
 
           # @rbs node: RuboCop::AST::SendNode
           def on_send: (RuboCop::AST::SendNode node) -> void
-
-          private
-
-          # @rbs node: RuboCop::AST::SendNode
-          def struct_like_class?: (RuboCop::AST::SendNode node) -> bool
         end
       end
     end
